minor changes to libft functions

+++ include/libft.h
@@ -234,7 +234,7 @@ char		*ft_strsub(char const *s, unsigned int start, size_t len);
 char		*ft_strjoin(char const *s1, char const *s2);
 char		*ft_strtrim(char const *s);
 char		**ft_strsplit(char const *s, char c);
-char		*ft_itoa(int n);
+char		*ft_itoa(intmax_t n);
 char		*ft_itoa_base(intmax_t value, int base);
 void		ft_putchar(char c);
 void		ft_putstr(char const *s);
